+++ front/components/spaces/MembersTable.tsx
@@ -1,0 +1,137 @@
+import { Button, DataTable, XMarkIcon } from "@dust-tt/sparkle";
+import type {
+  CellContext,
+  PaginationState,
+  SortingState,
+} from "@tanstack/react-table";
+import { useCallback, useMemo, useState } from "react";
+
+import { useSendNotification } from "@app/hooks/useNotification";
+import type { UserType } from "@app/types";
+
+type MemberRowData = {
+  email: string;
+  icon: string;
+  name: string;
+  onClick?: () => void;
+  userId: string;
+};
+
+type MemberInfo = CellContext<MemberRowData, unknown>;
+
+function getMemberTableRows(allUsers: UserType[]): MemberRowData[] {
+  return allUsers.map((user) => ({
+    icon: user.image ?? "",
+    name: user.fullName,
+    userId: user.sId,
+    email: user.email ?? "",
+  }));
+}
+
+interface MembersTableProps {
+  onMembersUpdated: (members: UserType[]) => void;
+  searchSelectedMembers: string;
+  selectedMembers: UserType[];
+}
+
+export function MembersTable({
+  onMembersUpdated,
+  searchSelectedMembers,
+  selectedMembers,
+}: MembersTableProps) {
+  const sendNotifications = useSendNotification();
+  const [pagination, setPagination] = useState<PaginationState>({
+    pageIndex: 0,
+    pageSize: 50,
+  });
+  const [sorting, setSorting] = useState<SortingState>([
+    { id: "email", desc: false },
+  ]);
+
+  const getTableColumns = useCallback(() => {
+    const removeMember = (userId: string) => {
+      if (selectedMembers.length === 1) {
+        sendNotifications({
+          title: "Cannot remove last member.",
+          description: "You cannot remove the last group member.",
+          type: "error",
+        });
+        return;
+      }
+      onMembersUpdated(selectedMembers.filter((m) => m.sId !== userId));
+    };
+    return [
+      {
+        id: "name",
+        accessorKey: "name",
+        cell: (info: MemberInfo) => (
+          <>
+            <DataTable.CellContent
+              avatarUrl={info.row.original.icon}
+              className="hidden md:flex"
+            >
+              {info.row.original.name}
+            </DataTable.CellContent>
+            <DataTable.CellContent
+              avatarUrl={info.row.original.icon}
+              className="flex md:hidden"
+              description={info.row.original.email}
+            >
+              {info.row.original.name}
+            </DataTable.CellContent>
+          </>
+        ),
+        enableSorting: true,
+      },
+      {
+        id: "email",
+        accessorKey: "email",
+        cell: (info: MemberInfo) => (
+          <DataTable.BasicCellContent label={info.row.original.email} />
+        ),
+        enableSorting: true,
+      },
+      {
+        id: "action",
+        meta: {
+          className: "w-12",
+        },
+        cell: (info: MemberInfo) => {
+          return (
+            <DataTable.CellContent>
+              <Button
+                icon={XMarkIcon}
+                size="xs"
+                variant="ghost-secondary"
+                onClick={() => removeMember(info.row.original.userId)}
+              />
+            </DataTable.CellContent>
+          );
+        },
+      },
+    ];
+  }, [onMembersUpdated, selectedMembers, sendNotifications]);
+
+  const rows = useMemo(
+    () => getMemberTableRows(selectedMembers),
+    [selectedMembers]
+  );
+  const columns = useMemo(() => getTableColumns(), [getTableColumns]);
+
+  return (
+    <DataTable
+      data={rows}
+      columns={columns}
+      columnsBreakpoints={{
+        name: "md",
+      }}
+      pagination={pagination}
+      setPagination={setPagination}
+      sorting={sorting}
+      setSorting={setSorting}
+      totalRowCount={rows.length}
+      filter={searchSelectedMembers}
+      filterColumn="email"
+    />
+  );
+}
